Fixed up problems caused from previous update that appear when compiling WITH_GL_PROFILE_COMPAT and WITH_GL_SYSTEM_LEGACY (as opposed to WITH_GL_PROFILE_ES20 and WITH_GL_SYSTEM_EMBEDDED)

+++ source/blender/blenlib/intern/math_vector_inline.c
@@ -609,7 +609,7 @@ MINLINE float len_manhattan_v2(const float v[2])
 
 MINLINE float len_manhattan_v2_int(const int v[2])
 {
-	return ABS(v[0]) + ABS(v[1]);
+	return (float)(ABS(v[0]) + ABS(v[1]));
 }
 
 MINLINE float len_manhattan_v3(const float v[3])
